separate single and multiple location charts
